fix package:adb tests
